refactor whole project's folders and files structure, create HEALTH_EMOJIS dict, refactor mock_api.py
