+++ bot/constants/node_ips.py
@@ -1,20 +1,11 @@
 import os
-import logging
 
-# Enable logging
-logging.basicConfig(
-    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
-    level=logging.INFO)
-logger = logging.getLogger(__name__)
-
-DEBUG = bool(os.environ['DEBUG'] == 'True') if 'DEBUG' in os.environ else False
-TELEGRAM_BOT_TOKEN = os.environ.get('TELEGRAM_BOT_TOKEN', '')
+from constants.globals import DEBUG, logger
 
 NETWORK_TYPES = ["TESTNET", "CHAOSNET"]
 NETWORK_TYPE = os.getenv("NETWORK_TYPE").upper() \
     if os.getenv("NETWORK_TYPE", "notFound").upper() in NETWORK_TYPES and not DEBUG else 'TESTNET'
 
-# Set BINANCE_NODE_IP depending on mode (if None, no Binance jobs are not executed)
 if DEBUG:
     BINANCE_NODE_IPS = ['localhost', '0.0.0.0']
     BINANCE_DEX_ENDPOINT = "https://testnet-dex-atlantic.binance.org"
@@ -39,39 +30,3 @@ if (len(BITCOIN_NODE_IPS) != len(BITCOIN_NODE_USERNAMES)) or (len(BITCOIN_NODE_I
                    f"BITCOIN_NODE_USERNAMES array length: ({len(BITCOIN_NODE_USERNAMES)})\n"
                    f"BITCOIN_NODE_PASSWORDS array length: ({len(BITCOIN_NODE_PASSWORDS)})\n")
     BITCOIN_NODE_IPS.clear()
-
-if os.getenv("ALLOWED_USER_IDS", "notFound").upper() == 'ALL':
-    ALLOWED_USER_IDS = 'ALL'
-elif os.getenv("ALLOWED_USER_IDS"):
-    ALLOWED_USER_IDS = [int(allowed_id) for allowed_id in os.getenv("ALLOWED_USER_IDS").split(",")]
-else:
-    ALLOWED_USER_IDS = []
-
-# By how much we multiply the notification timeout in case of continuous Thornode attribute changes
-NOTIFICATION_TIMEOUT_MULTIPLIER = 1.5
-# Base notification timeout in seconds
-INITIAL_NOTIFICATION_TIMEOUT = 15
-
-# Emojis for status of THORNodes (note that keys are displayed to the user)
-STATUS_EMOJIS = {
-    "unknown": "❓",
-    "whitelisted": "📋",
-    "standby": "📆",
-    "ready": "✔️",
-    "active": "💚",
-    "disabled": "🔴"
-}
-
-MONITORED_STATUSES = ["standby", "ready", "active"]
-
-JOB_INTERVAL_IN_SECONDS = 5 if DEBUG else 30
-
-# Paths
-storage_path = os.sep.join(
-    [os.path.dirname(os.path.realpath(__file__)), os.path.pardir, 'storage'])
-session_data_path = os.sep.join([storage_path, 'session.data'])
-
-CONNECTION_TIMEOUT = 10
-MISSING_FUNDS_THRESHOLD = 10  # Number of cycles that thorchain can be insolvent before a message is sent
-
-REQUEST_POSTFIX = '?height=0'  # currently needed to get correct results due to a bug in thornodes
